refactor: refactor getting endpoints
