+++ jina/serve/runtimes/head/request_handling.py
@@ -17,64 +17,10 @@ from jina.serve.runtimes.monitoring import MonitoringRequestMixin
 from jina.serve.runtimes.worker.request_handling import WorkerRequestHandler
 from jina.types.request.data import DataRequest, Response
 from jina._docarray import docarray_v2
-
 if docarray_v2:
-    from docarray.base_doc import AnyDoc
-    from jina._docarray import docarray_v2
-    from docarray import DocList, BaseDoc
-
-    def _create_pydantic_model_from_schema(schema: Dict[str, any], model_name: str) -> type:
-        from pydantic import create_model
-        fields = {}
-        for field_name, field_schema in schema.get('properties', {}).items():
-            field_type = field_schema.get('type', None)
-            if field_type == 'string':
-                field_type = str
-            elif field_type == 'integer':
-                field_type = int
-            elif field_type == 'number':
-                field_type = float
-            elif field_type == 'boolean':
-                field_type = bool
-            elif field_type == 'array':
-                field_item_type = field_schema.get('items', {}).get('type', None)
-                if field_item_type == 'string':
-                    field_type = List[str]
-                elif field_item_type == 'integer':
-                    field_type = List[int]
-                elif field_item_type == 'number':
-                    field_type = List[float]
-                elif field_item_type == 'boolean':
-                    field_type = List[bool]
-                elif field_item_type == 'object' or field_item_type is None:
-                    # Check if array items are references to definitions
-                    items_ref = field_schema.get('items', {}).get('$ref')
-                    if items_ref:
-                        ref_name = items_ref.split('/')[-1]
-                        field_type = DocList[_create_pydantic_model_from_schema(schema['definitions'][ref_name], ref_name)]
-                    else:
-                        field_type = DocList[_create_pydantic_model_from_schema(field_schema.get('items', {}), field_name)]
-                else:
-                    raise ValueError(f"Unknown array item type: {field_item_type} for field_name {field_name}")
-            elif field_type == 'object' or field_type is None:
-                # Check if object is a reference to definitions
-                if 'additionalProperties' in field_schema:
-                    additional_props = field_schema['additionalProperties']
-                    if additional_props.get('type') == 'object':
-                        field_type = Dict[str, _create_pydantic_model_from_schema(additional_props, field_name)]
-                    else:
-                        field_type = Dict[str, Any]
-                else:
-                    obj_ref = field_schema.get('$ref')
-                    if obj_ref:
-                        ref_name = obj_ref.split('/')[-1]
-                        field_type = _create_pydantic_model_from_schema(schema['definitions'][ref_name], ref_name)
-                    else:
-                        field_type = _create_pydantic_model_from_schema(field_schema, field_name)
-            else:
-                raise ValueError(f"Unknown field type: {field_type} for field_name {field_name}")
-            fields[field_name] = (field_type, field_schema.get('description'))
-        return create_model(model_name, __base__=BaseDoc, **fields)
+    from jina.serve.runtimes.helper import _create_pydantic_model_from_schema
+    from docarray import DocList
+    from docarray.base_doc.any_doc import AnyDoc
 
 
 if TYPE_CHECKING:  # pragma: no cover
